Fix leader addIndexer
