+++ pkg/controller/master/setup.go
@@ -3,15 +3,48 @@ package master
 import (
 	"context"
 
-	"github.com/rancher/harvester/pkg/config"
 	"github.com/rancher/steve/pkg/server"
 	"github.com/rancher/wrangler/pkg/leader"
+
+	"github.com/rancher/harvester/pkg/config"
+	"github.com/rancher/harvester/pkg/controller/master/auth"
+	"github.com/rancher/harvester/pkg/controller/master/image"
+	"github.com/rancher/harvester/pkg/controller/master/keypair"
+	"github.com/rancher/harvester/pkg/controller/master/node"
+	"github.com/rancher/harvester/pkg/controller/master/template"
+	"github.com/rancher/harvester/pkg/controller/master/user"
+	"github.com/rancher/harvester/pkg/controller/master/virtualmachine"
+	"github.com/rancher/harvester/pkg/indexeres"
 )
+
+type registerFunc func(context.Context, *config.Management) error
+
+var registerFuncs = []registerFunc{
+	image.Register,
+	keypair.Register,
+	node.Register,
+	template.Register,
+	virtualmachine.Register,
+	user.Register,
+}
+
+func register(ctx context.Context, management *config.Management) error {
+	for _, f := range registerFuncs {
+		if err := f(ctx, management); err != nil {
+			return err
+		}
+	}
+
+	return auth.BootstrapAdmin(management)
+}
 
 func Setup(ctx context.Context, server *server.Server) error {
 	scaled := config.ScaledWithContext(ctx)
+
+	indexeres.RegisterManagementIndexers(scaled.Management)
+
 	go leader.RunOrDie(ctx, "", "harvester-controllers", server.K8s, func(ctx context.Context) {
-		if err := register(ctx, server, scaled.Management); err != nil {
+		if err := register(ctx, scaled.Management); err != nil {
 			panic(err)
 		}
 		if err := scaled.Management.Start(); err != nil {
